+++ changelog.md
@@ -1,5 +1,8 @@
+### 9/13/2023 0.4.0-alpha.3
+- fix issue of not picking up shadowgenerator when loading saved scene
 ### 9/13/2023 0.4.0-alpha.2
 - fix issue of unicam not moving when Cascade shadow map used
+- fix issue of not picking up shadowgenerator when loading saved scene
 - unicam speed increases up when shft key is pressed
 ### 9/13/2023 0.4.0-alpha.1
 - azimuth and elevation algorithm changed